backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Optional: enable OpenAI insights if OPENAI_API_KEY is set
openai==1.56.0
